requirements.txt
# Core data & ML
pandas>=2.2
numpy>=1.26
scipy>=1.11
scikit-learn>=1.4

# Embeddings / NLP
sentence-transformers>=2.5
torch>=2.2
sentencepiece>=0.1.99

# Clustering / dimensionality reduction
hdbscan>=0.8.33
umap-learn>=0.5.6

# LLM tooling & providers
transformers>=4.43
huggingface-hub>=0.23
accelerate>=0.31
google-generativeai>=0.7

# CLI, config, utils
typer>=0.12
python-dotenv>=1.0.1
pydantic>=2.7
requests>=2.32
tqdm>=4.66
rich>=13.7
pyyaml>=6.0.1

# Viz
matplotlib>=3.8
plotly>=5.22

# Explainability (for classical baselines)
shap>=0.45
lime>=0.2.0.1

# Optional topic modeling / exploration
# bertopic>=0.16

# Optional dashboards
# streamlit>=1.36
